CBD-3947: Update Server base image to be based on UBI 8
Also includes some more Dockerfile cleanups, plus builds runit from
source since there are no RHEL 8 packages available

+++ scripts/build-rhcc-image.sh
@@ -6,17 +6,20 @@ function show_help {
     echo "Options:"
     echo "  -p : Product to build (e.g. couchbase-server) (Required)"
     echo "  -v : Version of product to use (e.g. 5.1.0) (Required)"
-    echo "  -t : Build test image incl. sshd (Optional, defaults to false)"
     echo "  -s : Build from staging repository (Optional, defaults to false)"
+    echo "  -q : Quick build (disables docker build --no-cache)"
+    echo "  -n : Dry run (don't push to gitlab)"
     exit 0
 }
 
 SCRIPT_DIR=$(dirname $(readlink -e -- "${BASH_SOURCE}"))
 
 STAGING=""
+CACHE_ARG="--no-cache"
+DRYRUN=""
 
 # Parse options and ensure required ones are there
-while getopts :p:v:b:tsh opt; do
+while getopts :p:v:b:qsnh opt; do
     case ${opt} in
         p) PRODUCT="$OPTARG"
            ;;
@@ -24,9 +27,11 @@ while getopts :p:v:b:tsh opt; do
            ;;
         b) BUILD="$OPTARG"
            ;;
-        t) TESTING="true"
-           ;;
         s) STAGING="-staging"
+           ;;
+        q) CACHE_ARG=""
+           ;;
+        n) DRYRUN="yes"
            ;;
         h) show_help
            ;;
@@ -50,12 +55,19 @@ if [[ -z "$BUILD" ]]; then
     BUILD=1
 fi
 
+# Use new UBI-based Dockerfile for Server 7.x or later, or SGW 3.x or later
+if [[ ${PRODUCT} == "couchbase-server" && ${VERSION} =~ 6.* ]]; then
+    echo "Using old-school Dockerfile.old"
+    DOCKERFILE=Dockerfile.old
+elif [[ ${PRODUCT} == "sync-gateway" && ${VERSION} =~ 2.* ]]; then
+    echo "Using old-school Dockerfile.old"
+    DOCKERFILE=Dockerfile.old
+else
+    DOCKERFILE=Dockerfile
+fi
+
 # Enter product directory
 cd ${PRODUCT}
-
-# Ensure base image is up-to-date
-BASE_IMAGE=$(grep '^FROM' Dockerfile|cut -d' ' -f2)
-docker pull ${BASE_IMAGE}
 
 # Some informational settings
 CONF_DIR=/home/couchbase/openshift/${PRODUCT}
@@ -65,14 +77,15 @@ IMAGE=${INTERNAL_IMAGE_NAME}:${VERSION}-${BUILD}
 
 # Build image
 ${SCRIPT_DIR}/update-base.sh Dockerfile
-docker build --no-cache \
+docker build ${CACHE_ARG} \
   --build-arg PROD_VERSION=${VERSION} \
   --build-arg STAGING=${STAGING} \
-  --build-arg TESTING=${TESTING} \
-  -f Dockerfile -t ${IMAGE} .
+  -f ${DOCKERFILE} -t ${IMAGE} .
 
-echo @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
-echo Pushing ${IMAGE}
-echo @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
-docker push ${IMAGE}
-docker rmi ${IMAGE}
+if [ "x${DRYRUN}" != "xyes" ]; then
+    echo @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
+    echo Pushing ${IMAGE}
+    echo @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
+    docker push ${IMAGE}
+    docker rmi ${IMAGE}
+fi
